refactor: allow for generalized timeout/retry logic
A fairly substantial refactor that redoes some of the logic, and
especially the means of communicating to the user.

This currently just applies hardcoded settings for both. The next step
will be to change the config format to allow setting them on a
per-check basis.

After that I'll want to clean up the output, perhaps to something more
like Colmena itself.