feat: replace --on with --select and a selection DSL
This adds labels to each check definition, and a selection DSL to use
them.

Right now the DSL is relatively simple, allowing selection by a single
label name, and either a list of exact strings, or a regex.

In the future I want to add grouping, negation, and and/or semantics
at least. For now, this is sufficient to replace --on.