refactor: move one-off checking to separate function
This moves the current behavior out to a separate function, leaving
space to implement alerts using a common interface.

Some of this should also make it easier to implement a more concise
mode of output for reports. Currently this means that all output is
effectively verbose-mode, but that will be fixed in a future commit.