refactor: require all configuration fields
I realized while implementing alerts that I let too much complication
come in with all of the multiple levels of defaulting. The idea was to
make it easier to write the configs, but ultimately they're supposed
to be generated anyway.

This pushes the defaulting behavior to the Nix-side, where it's much
more natural. It also means that the check configurations come in
ready to be used directly.